added login seeker page and renamed register components
