+++ client/src/components/index.js
@@ -1,11 +1,16 @@
 
 import EmployerLogin from './employer/EmployerLogin';
 import EmployerProfile from './employer/EmployerProfile';
-import RegisterEmployer from './employer/RegisterEmployer';
-import RegisterSeeker from './seeker/RegisterSeeker';
+import EmployerRegister from './employer/EmployerRegister';
+import SeekerRegister from './seeker/SeekerRegister';
 import EmployerBrowseMatches from './employer/EmployerBrowseMatches';
+import SeekerLogin from './seeker/SeekerLogin';
 
 export {
-  RegisterEmployer, RegisterSeeker, EmployerLogin,
-  EmployerProfile, EmployerBrowseMatches,
+  EmployerRegister,
+  SeekerRegister,
+  EmployerLogin,
+  EmployerProfile,
+  EmployerBrowseMatches,
+  SeekerLogin,
 };
