+++ server/server.js
@@ -15,11 +15,11 @@ const jobRouter = require('./data/jobs/jobRoutes');
 const server = express();
 
 const corsOptions = {
-  origin: ('http://localhost:3000', 'https://jobitduder.herokuapp.com'),
-    credentials: true,
-    methods: ['GET', 'PUT', 'POST'],
-    allowedHeaders: ['Content-Type', 'Authorization']
-}
+  origin: ['http://localhost:3000', 'https://jobitduder.herokuapp.com'],
+  credentials: true,
+  methods: ['GET', 'PUT', 'POST'],
+  allowedHeaders: ['Content-Type', 'Authorization'],
+};
 server.use(morgan());
 server.use(express.json());
 server.use(cors(corsOptions));
